Skip duplicate contributions
If easytithe_importer.py is re-ran, it will now skip any duplicate
entries already added to Breeze. The script will look for contributions
given for a person on a specific date and if there’s matching
contribution, it will skip adding it.

This isn’t a fool-proof algorithm: congregants may actually contribute
the same amounts twice on a day. I may add a flag to forcibly add
duplicate contributions in the future.