Upgrade duplicates loggging from INFO to WARNING
When a duplicate contribution amount is found, log it as a warning.
This way, I can set an alert in our Heroku Scheduler that will fire as
soon as it occurs so I can investigate.

There could actually be a legitimate duplicate contribution submitted
by the same person on a given day and I want to know about it as soon
as it happens so I can make sure it gets imported into Breeze.